ai-garbage-child/requirements.txt
requests
fastapi
uvicorn
pydantic
python-multipart
pillow
numpy

